rename file and tidy up import for clarity, bugfix
