+++ backend/server/routes/establishments/bulkUpload/bulkUploadS3Client.js
@@ -2,6 +2,6 @@ const config = require('../../../config/config');
 const { S3ClientV3 } = require('../../../aws/s3Client');
 
 const region = String(config.get('bulkupload.region'));
-const s3Client = new S3ClientV3(region);
+const bulkUploadS3Client = new S3ClientV3(region);
 
-module.exports = s3Client;
+module.exports = bulkUploadS3Client;
